ACTUALIZACIÓN: Se añadieron migraciones, vistas, templates, formularios y rutas para la gestión de movimientos de stock, incluyendo filtros y operaciones como transferencias, ajustes, entradas y salidas. Se realizaron ajustes en modelos, validación de stock, y modales para operaciones contextuales.

+++ bodega/models/movimientos.py
@@ -29,10 +29,7 @@ class MovimientoInventario(StandardModelEstablishment):
         choices=TIPO_CHOICES
     )
 
-    cantidad = models.DecimalField(
-        max_digits=12,
-        decimal_places=2
-    )
+    cantidad = models.PositiveIntegerField()
 
     usuario = models.ForeignKey(
         settings.AUTH_USER_MODEL,
